requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
requests
gunicorn
